+++ jason/src/api/connection.rs
@@ -62,7 +62,7 @@ impl Connection {
     /// Invoke `on_remote_stream` [`Connection`]'s callback
     /// for a given [`MediaStream`] received from a related remote [`Member`].
     #[inline]
-    pub(crate) fn on_remote_stream(&self, stream_handle: MediaStreamHandle) {
-        self.0.borrow().on_remote_stream.call(stream_handle);
+    pub(crate) fn on_remote_stream(&self, stream: MediaStreamHandle) {
+        self.0.borrow().on_remote_stream.call(stream);
     }
 }