+++ jason/CHANGELOG.md
@@ -14,7 +14,7 @@ All user visible changes to this project will be documented in this file. This p
 ### BC Breaks
 
 - Library API:
-    - Bind `on_local_stream` callback to `Room` instead of `Jason` ([#54]);
+    - Expose `on_local_stream` callback in `Room` instead of `Jason` ([#54]);
     - Remove error argument from `on_local_stream` callback ([#54]);
     - Room initialization ([#46]):
         - Remove `Jason.join_room()`.
@@ -36,8 +36,8 @@ All user visible changes to this project will be documented in this file. This p
         - Room initialization ([#46]):
             - `Jason.init_room()`;
             - `Room.join()`;
-        - Inject local video/audio into `Room` via `Room.inject_local_stream()` ([#54]);
-        - Add `Room.on_failed_local_stream` callback ([#54]).
+        - Ability to inject local video/audio stream into `Room` via `Room.inject_local_stream()` ([#54]);
+        - `Room.on_failed_local_stream` callback ([#54]).
 
 ### Fixed